% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Goodbye Function}
\usage{
goodbye(my_name)
}
\arguments{
\item{my_name}{Name to say goodbye to.}
}
\value{
The printed string.
}
\description{
Prints a goodbye message for a supplied name.
}
\examples{
goodbye("Ben")
}
