% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello Function}
\usage{
hello(my_name)
}
\arguments{
\item{my_name}{name to say hello to.}
}
\value{
The printed string.
}
\description{
This function just prints "Hello, <Name>" for fun.
Daniel Sabanes wrote this function.
}
\examples{
hello("Ben")
}
